+++ moveit_core/single_waypt_smoothing_plugins/CMakeLists.txt
