+++ moveit_core/single_waypt_smoothing_plugins/include/moveit/single_waypt_smoothing_plugins/butterworth_filter.h
@@ -41,9 +41,9 @@
 #include <cstddef>
 
 #include <moveit/robot_model/robot_model.h>
-#include <moveit/smoothing_plugins/smoothing_base_class.h>
+#include <moveit/single_waypt_smoothing_plugins/smoothing_base_class.h>
 
-namespace smoothing_plugins
+namespace single_waypt_smoothing_plugins
 {
 /**
  * Class ButterworthFilter - Implementation of a signal filter to soften jerks.
@@ -110,4 +110,4 @@ private:
   std::vector<ButterworthFilter> position_filters_;
   size_t num_joints_;
 };
-}  // namespace smoothing_plugins
+}  // namespace single_waypt_smoothing_plugins
