+++ moveit_ros/moveit_servo/test/config/servo_settings_low_latency.yaml
@@ -27,7 +27,7 @@ publish_joint_velocities: false
 publish_joint_accelerations: false
 
 ## Plugins for smoothing outgoing commands
-smoothing_filter_plugin_name: "smoothing_plugins::ButterworthFilterPlugin"
+smoothing_filter_plugin_name: "single_waypt_smoothing_plugins::ButterworthFilterPlugin"
 
 ## MoveIt properties
 move_group_name:  panda_arm  # Often 'manipulator' or 'arm'
